I didn't even notice I was casting y to the wrong type in the _rotateright64() builtin wrapper, but the compiler apparently didn't mind anyway. It only seems to care that the builtin itself is either inferred or declared with the correct type. As such, I just got rid of the cast (casting a type to itself is a bit silly anyway), and it's really just a shame in the end that one missing letter makes using the macro to make this declaration no longer worth the trouble. It's still a ridiculous bug to have shown up in the first place.

+++ Sources/CDiscretionalClangBuiltins/include/CDiscretionalClangBuiltins.h
@@ -45,7 +45,7 @@ BitwiseRotationBuiltin(rotateright32, ror, uint32_t)
 // from Xcode 11.4) does not have this fix yet. Hopefully the next release of Xcode will have it.
 //BitwiseRotationBuiltin(rotateright64, rorll, uint64_t)
 __attribute__((swift_name("ClangBuiltins.rorll(x:y:)")))
-static inline uint64_t _rotateright64(uint64_t x, uint64_t y) { return __builtin_rotateright64(x, (uint64_t)y); }
+static inline uint64_t _rotateright64(uint64_t x, uint64_t y) { return __builtin_rotateright64(x, y); }
 extern uint64_t __attribute__((availability(swift,unavailable))) __builtin_rotateright64(uint64_t, int64_t);
 
 #undef BitwiseRotationBuiltin
